requirements.txt
pandas==1.3.5
flask==1.1.4
flask-cors==3.0.10
flask-marshmallow
flask-oidc
flask-script
flask-migrate
flask-sqlalchemy==2.5.1
flask-login
sqlalchemy==1.4.23
marshmallow
marshmallow-sqlalchemy
psycopg2==2.9.1
python-dotenv==0.19.0
werkzeug==1.0.1
openpyexcel
requests==2.26.0
simplejson==3.17.5
markupsafe==2.0.1